+++ ElvUI_MerathilisUI/Locales/esMX.lua
@@ -475,6 +475,7 @@ L["Alpha Max"] = true
 L["Inherit Global Fade"] = true
 L["Anchor Point"] = true
 L["The first button anchors itself to this point on the bar."] = true
+L["Dream Seeds"] = true
 
 -- Media
 L["Zone Text"] = true
@@ -725,6 +726,7 @@ L["Placeholders"] = true
 L["Use Default"] = true
 L["Researchers Under Fire"] = "Investigadores bajo el fuego"
 L["Time Rift Thaldraszus"] = "Falla temporal Thaldraszus"
+L["Superbloom"] = true
 
 -- SMB
 L["Minimap Buttons"] = true