+++ ElvUI_MerathilisUI/Locales/frFR.lua
@@ -490,6 +490,7 @@ L["Alpha Max"] = "Alpha Max"
 L["Inherit Global Fade"] = true
 L["Anchor Point"] = true
 L["The first button anchors itself to this point on the bar."] = true
+L["Dream Seeds"] = true
 
 -- Media
 L["Zone Text"] = true
@@ -740,6 +741,7 @@ L["Placeholders"] = true
 L["Use Default"] = true
 L["Researchers Under Fire"] = "Des chercheurs sous le feu"
 L["Time Rift Thaldraszus"] = "Faille dans le temps Thaldraszus"
+L["Superbloom"] = true
 
 -- SMB
 L["Minimap Buttons"] = "Bouton de la minicarte"