+++ ElvUI_MerathilisUI/Locales/zhCN.lua
@@ -493,6 +493,7 @@ L["Alpha Max"] = "最大透明度"
 L["Inherit Global Fade"] = true
 L["Anchor Point"] = true
 L["The first button anchors itself to this point on the bar."] = true
+L["Dream Seeds"] = true
 
 -- Media
 L["Zone Text"] = "区域文字"
@@ -746,6 +747,7 @@ L["Placeholders"] = "占位符"
 L["Use Default"] = "使用默认"
 L["Researchers Under Fire"] = true
 L["Time Rift Thaldraszus"] = true
+L["Superbloom"] = true
 
 -- SMB
 L["Minimap Buttons"] = "小地图按钮"