+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ### Changes ###
 
-  * Fixed an oopsie in the BagSync Skin
+  * Added Superbloom Timer on the Event Tracker
+  * Added a check on the Mail Module to update Alts on login
+  * Added new Openables to the AutoButtons. thx @mcc1
+  * Take account of some new Font Settings from ElvUI
+  * Removed the Media Section, cause now in stock ElvUI
+  * Fixed some issues on the Chat Text
+  * Fixed shadow on my AFK-Screen